requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
PyMuPDF==1.23.26
google-generativeai==0.5.4
pyautogen==0.2.20
